refactor: extract MCP normalization tests to stay under 300-line limit
Split buildServerOptions MCP type normalization tests into dedicated
tests/mcp-normalization.test.js (167 lines) to bring
tests/opencode-client-cowork.test.js back under the 300-line limit
(169 lines). No test logic changed.
